requirements.txt
jinja2>=3.1.0
PyYAML>=6.0
requests>=2.28.0
matplotlib>=3.7.0
numpy>=1.24.0
anthropic>=0.25.0
